Extract filtering logic into a standalone hook
